refactor specs
