+++ card/mod/account/spec/set/all/account_spec.rb
@@ -1,139 +1,132 @@
 # -*- encoding : utf-8 -*-
 
-describe Card::Set::All::Account do
+RSpec.describe Card::Set::All::Account do
   describe "accountable?" do
     it "is false for cards with *accountable rule off" do
       expect(Card["A"].accountable?).to eq(false)
     end
 
-    it "is true for cards with *accountable rule on" do
-      Card::Auth.as_bot do
-        Card.create name: "A+*self+*accountable", content: "1"
-        Card.create name: "*account+*right+*create",
-                    content: "[[Anyone Signed In]]"
-      end
+    it "is true for cards with *accountable rule on", as_bot: true do
+      Card.create name: "A+*self+*accountable", content: "1"
+      Card.create name: "*account+*right+*create",
+                  content: "[[Anyone Signed In]]"
       expect(Card["A"].accountable?).to eq(true)
     end
-  end
 
-  describe "parties" do
-    it "for Wagn Bot" do
-      Card::Auth.current_id = Card::WagnBotID
-      expect(Card::Auth.current.parties.sort).to eq(
-        [Card::WagnBotID, Card::AnyoneSignedInID, Card::AdministratorID]
-      )
-    end
-
-    it "for Anonymous" do
-      Card::Auth.current_id = Card::AnonymousID
-      expect(Card::Auth.current.parties.sort).to eq([Card::AnonymousID])
-    end
-
-    context "for Joe User" do
-      before do
-        @joe_user_card = Card::Auth.current
-        @parties = @joe_user_card.parties # note: must be called to test resets
+    describe "parties" do
+      it "for Wagn Bot", with_user: Card::WagnBotID do
+        expect(Card::Auth.current.parties.sort).to eq(
+                                                     [Card::WagnBotID, Card::AnyoneSignedInID, Card::AdministratorID]
+                                                   )
       end
 
-      it "initially has only auth and self " do
-        expect(@parties).to eq([Card::AnyoneSignedInID, @joe_user_card.id])
+      it "for Anonymous", with_user: Card::AnonymousID do
+        expect(Card::Auth.current.parties.sort).to eq([Card::AnonymousID])
       end
 
-      it "updates when new roles are set" do
-        roles_card = @joe_user_card.fetch trait: :roles, new: {}
-        r1 = Card["r1"]
+      context "for Joe User" do
+        before do
+          @joe_user_card = Card::Auth.current
+          @parties = @joe_user_card.parties # note: must be called to test resets
+        end
 
-        Card::Auth.as_bot { roles_card.items = [r1.id] }
-        expect(Card["Joe User"].parties).to eq(@parties)
-        # local cache still has old parties
-        # (permission does not change mid-request)
+        it "initially has only auth and self " do
+          expect(@parties).to eq([Card::AnyoneSignedInID, @joe_user_card.id])
+        end
 
-        Card::Cache.restore
-        # simulate new request
-        # clears local cache, where, eg, @parties would still be cached on card
+        it "updates when new roles are set" do
+          roles_card = @joe_user_card.fetch trait: :roles, new: {}
+          r1 = Card["r1"]
 
-        Card::Auth.current_id = Card::Auth.current_id
-        # simulate new request
-        # current_id assignment clears several class variables
+          Card::Auth.as_bot { roles_card.items = [r1.id] }
+          expect(Card["Joe User"].parties).to eq(@parties)
+          # local cache still has old parties
+          # (permission does not change mid-request)
 
-        new_parties = [Card::AnyoneSignedInID, r1.id, @joe_user_card.id]
-        expect(Card["Joe User"].parties).to eq(new_parties)
-        # @parties regenerated, now with correct values
+          Card::Cache.restore
+          # simulate new request
+          # clears local cache, where, eg, @parties would still be cached on card
 
-        expect(Card::Auth.current. parties).to eq(new_parties)
-        # @joe_user_card.refresh(force=true).parties.should == new_parties
-        # should work, but now superfluous?
-      end
-    end
-  end
+          Card::Auth.current_id = Card::Auth.current_id
+          # simulate new request
+          # current_id assignment clears several class variables
 
-  describe "among?" do
-    it "is true for self" do
-      expect(Card::Auth.current.among?([Card::Auth.current_id])).to be_truthy
-    end
-  end
+          new_parties = [Card::AnyoneSignedInID, r1.id, @joe_user_card.id]
+          expect(Card["Joe User"].parties).to eq(new_parties)
+          # @parties regenerated, now with correct values
 
-  describe "+*email" do
-    it "creates a card and account card" do
-      jadmin = Card["joe admin"]
-      Card::Auth.current_id = jadmin.id
-      # simulate login to get correct from address
-
-      Card::Env[:params] = { email: { subject: "Hey Joe!",
-                                      message: "Come on in." } }
-      Card.create! name: "Joe New",
-                   type_id: Card::UserID,
-                   "+*account" => { "+*email" => "joe@new.com" }
-
-      c = Card["Joe New"]
-      u = Card::Auth.find_account_by_email("joe@new.com")
-
-      expect(c.account).to eq(u)
-      expect(c.type_id).to eq(Card::UserID)
-    end
-  end
-
-  context "updates" do
-    before do
-      @card = Card["Joe User"]
-    end
-
-    it "handles email updates" do
-      @card.update_attributes! "+*account" => { "+*email" => "joe@user.co.uk" }
-      expect(@card.account.email).to eq("joe@user.co.uk")
-    end
-
-    it "lets Wagn Bot block accounts" do
-      Card::Auth.as_bot do
-        @card.account.status_card.update_attributes! content: "blocked"
-        expect(@card.account.blocked?).to be_truthy
+          expect(Card::Auth.current.parties).to eq(new_parties)
+          # @joe_user_card.refresh(force=true).parties.should == new_parties
+          # should work, but now superfluous?
+        end
       end
     end
 
-    it "does not allow a user to block or unblock himself" do
-      expect do
-        @card.account.status_card.update_attributes! content: "blocked"
-      end.to raise_error(ActiveRecord::RecordInvalid,
-                         "Validation failed: Permission denied You don't have "\
+    describe "among?" do
+      it "is true for self" do
+        expect(Card::Auth.current.among?([Card::Auth.current_id])).to be_truthy
+      end
+    end
+
+    describe "+*email" do
+      it "creates a card and account card" do
+        jadmin = Card["joe admin"]
+        Card::Auth.current_id = jadmin.id
+        # simulate login to get correct from address
+
+        Card::Env[:params] = { email: { subject: "Hey Joe!",
+                                        message: "Come on in." } }
+        Card.create! name: "Joe New",
+                     type_id: Card::UserID,
+                     "+*account" => { "+*email" => "joe@new.com" }
+
+        c = Card["Joe New"]
+        u = Card::Auth.find_account_by_email("joe@new.com")
+
+        expect(c.account).to eq(u)
+        expect(c.type_id).to eq(Card::UserID)
+      end
+    end
+
+    context "updates" do
+      let(:card) {Card["Joe User"]}
+      let(:account) {card.account}
+
+      it "handles email updates" do
+        card.update_attributes! "+*account" => { "+*email" => "joe@user.co.uk" }
+        expect(account.email).to eq("joe@user.co.uk")
+      end
+
+      it "lets Wagn Bot block accounts", as_bot: true do
+        card.account.status_card.update_attributes! content: "blocked"
+        expect(account.blocked?).to be_truthy
+      end
+
+      it "does not allow a user to block or unblock himself" do
+        expect do
+          account.status_card.update_attributes! content: "blocked"
+        end.to raise_error(ActiveRecord::RecordInvalid,
+                           "Validation failed: Permission denied You don't have "\
                          "permission to change the status of your own account")
-      expect(@card.account.blocked?).to be_falsey
-    end
-  end
-
-  describe "#read_rules" do
-    before(:all) do
-      @read_rules = Card["joe_user"].read_rules
+        expect(account.blocked?).to be_falsey
+      end
     end
 
-    it "*all+*read should apply to Joe User" do
-      expect(@read_rules.member?(Card.fetch("*all+*read").id)).to be_truthy
-    end
+    describe "#read_rules" do
+      before(:all) do
+        @read_rules = Card["joe_user"].read_rules
+      end
 
-    it "14 more should apply to Joe Admin" do
-      # includes lots of account rules...
-      Card::Auth.as("joe_admin") do
-        ids = Card::Auth.as_card.read_rules
-        expect(ids.length).to eq(@read_rules.size + 14)
+      it "*all+*read should apply to Joe User" do
+        expect(@read_rules.member?(Card.fetch("*all+*read").id)).to be_truthy
+      end
+
+      it "14 more should apply to Joe Admin" do
+        # includes lots of account rules...
+        Card::Auth.as("joe_admin") do
+          ids = Card::Auth.as_card.read_rules
+          expect(ids.length).to eq(@read_rules.size + 14)
+        end
       end
     end
   end
