+++ card/spec/support/expectations.rb
@@ -1,0 +1,5 @@
+module CardExpectations
+  def expect_content_of name
+    expect(Card.fetch(name))
+  end
+end
